requirements.txt
certifi==2022.5.18.1
click==8.1.3
# dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib_1652035107515/work
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.2
Flask-Cors==3.0.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
